Create winapi_error module
This module exposes an `Error` type implementing the `std::error::Error`
trait, and a public `last()` function which returns the last `Error`
that occurred.

This abstracts over the logic to retrieve a human-readable error message
from Windows API errors, while also providing a `std::error::Error`
implementation that can be used as the return value for the
`source(&self)` method of higher-level `std::error::Error`s.